Update progress flutter side stream to double

+++ lib/providers/demixing_provider.dart
@@ -12,16 +12,16 @@ import 'package:get/route_manager.dart';
 class DemixingProvider extends ChangeNotifier {
   final _helper = DemixingHelper();
   final PreferencesProvider preferences;
-  late Stream<int> _progressStream;
+  late Stream<double> _progressStream;
   CancelableOperation<UnmixedSong>? _operation;
 
   DemixingProvider(this.preferences);
 
-  Stream<int> get progressStream => _progressStream;
+  Stream<double> get progressStream => _progressStream;
 
   void unmix(Song song, LibraryProvider library) {
     _progressStream =
-        _helper.progressStream.receiveBroadcastStream().cast<int>();
+        _helper.progressStream.receiveBroadcastStream().cast<double>();
     separate(song)
         ?.then((unmixed) => library.saveSong(unmixed))
         .then((index) => library.setCurrentSongIndex(index))
